backend/requirements.txt
# Flask Backend Requirements for Data Analysis Lab
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
statsmodels==0.14.1
scikit-learn==1.3.2
openpyxl==3.1.2
xlrd==2.0.1
pyreadstat==1.2.4
python-docx==1.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

